#19 refactor: 질문 상세 공통화 및 sessions 경로 통일
