+++ src/shared/pages/TestPage.tsx
@@ -20,7 +20,7 @@ function TestPage() {
 
   const testTabItems = [
     { label: "자료", path: "/test" },
-    { label: "과제", path: "/user/session" },
+    { label: "과제", path: "/user/sessions" },
     { label: "사용자 및 그룹", path: "/user/dashboard" },
     { label: "질문 및 답변", path: "/admin/sessions" },
   ];
