+++ src/user/domains/question/pages/UserQuestionsPage.tsx
@@ -3,6 +3,7 @@ import UserTitleSection from "@/user/shared/components/UserTitleSection";
 import QuestionTableHeader from "../components/QuestionTableHeader";
 import QuestionTableRows from "../components/QuestionTableRows";
 import type { SessionQuestionRow } from "../../session/types/SessionQuestionRow";
+// import TableEmptyState from "@/shared/components/table/TableEmptyState";
 
 const mockQuestions: { content: SessionQuestionRow[]; totalElements: number } =
   {
@@ -132,7 +133,7 @@ const mockQuestions: { content: SessionQuestionRow[]; totalElements: number } =
   };
 
 function UserQuestionsPage() {
-  const isLoading = false;
+  const isLoading = true;
 
   return (
     <div className="text-ec-black flex w-full max-w-251 flex-col gap-5 pt-7 pb-120 pl-30">
@@ -150,6 +151,8 @@ function UserQuestionsPage() {
         <div className="bg-ec-table-header rounded-tl-ec-10 rounded-tr-ec-10 flex items-center justify-between px-8 py-4">
           <QuestionTableHeader />
         </div>
+        {/* <TableEmptyState label="등록된 질문을 찾을 수 없거나 존재하지 않아요" /> */}
+
         <QuestionTableRows
           isLoading={isLoading}
           questions={mockQuestions.content}
