+++ src/user/layouts/SessionLayout.tsx
@@ -3,9 +3,9 @@ import { Outlet } from "react-router-dom";
 
 const testTabItems = [
   { label: "자료", path: "/test" },
-  { label: "과제", path: "/user/sessions" },
+  { label: "과제", path: "/user/sessions", end: true },
   { label: "사용자 및 그룹", path: "/user/dashboard" },
-  { label: "질문 및 답변", path: "/user/session/questions" },
+  { label: "질문 및 답변", path: "/user/sessions/questions" },
 ];
 
 function SessionLayout() {
